Added Careers microsite page and method for stacked tab widget
The framework now has a careers microsit page.  Only one page step def
class will represent all of the pages within the careers microsite.
Added a test for the stacked tab widget that checks if clicking on tab
opens the correct text field.  I can elaborate further on the assertion
to check if text is displayed in field?