Events Page: check if all articles contain search term
There is an issue here where it will check the article from search page,
assert search term.  If search term is not present from search page it
will open the article.  It then checks the main article descriptio for
search term.  Promlem is sometimes it will fail even the the article
contains term?