JobListingPage: Automated closing dates and click apply link
The closing date is difficult to assert because of the formate of
closing date.
The apply link works but could do with some work on adding if statement
if there are no job listings.  Also, maybe a better assertion for
checking the posting takes you to correct page rather than a partial
match of the url